:recycle: Little refactor. Methods and unused imports
