+++ src/components/Login/Login.tsx
@@ -1,10 +1,23 @@
 import React from 'react'
+import { StyledFirebaseAuth } from 'react-firebaseui';
+import firebase from 'firebase';
+
+const uiConfig = {
+  signInFlow: 'popup',
+  signInOptions: [
+    firebase.auth.GoogleAuthProvider.PROVIDER_ID
+  ],
+  callbacks: {
+    signInSuccessWithAuthResult: () => false
+  }
+}
+
 
 export const Login:React.FC = props => {
   return (
     <div className='login-container'>
       <h1>Login</h1>
-      
+      <StyledFirebaseAuth uiConfig={uiConfig} firebaseAuth={firebase.auth()} />
     </div>
   )
 }
